fix failing test due to import
