+++ tests/fixture_plugins/analysis_starter/name_fixtures.py
@@ -1,0 +1,7 @@
+import pytest
+
+
+@pytest.fixture
+def nextflow_case_id() -> str:
+    """Fixture for a Nextflow case id."""
+    return "case_id"
